+++ src/Mod/Draft/DraftGui.py
@@ -2452,6 +2452,3 @@ class ShapeStringTaskPanel:
 if not hasattr(FreeCADGui,"draftToolBar"):
     FreeCADGui.draftToolBar = DraftToolBar()
 #----End of Python Features Definitions----#
-
-if not hasattr(FreeCADGui,"Snapper"):
-    import DraftSnap
