+++ src/Mod/Draft/draftguitools/gui_snapper.py
